requirements.txt
asgiref==3.8.1
boto3==1.37.4
botocore==1.37.4
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-storages==1.14.5
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
